pyproject.toml
[tool.poetry]
name = "apis_in_a_nutshell"
version = "0.1.0"
description = ""
authors = ["Enrique Galindo <egalindo@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
